move useWindowWidth to hooks folder
